# docs/openapi.yaml
basePath: /api/v1
definitions:
...
  github_com_uigraph_app_internal_httputil.errorBody:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
info:
  contact: {}
  description: REST API for uigraph.
  title: uigraph API
  version: "1.0"
paths:
...
  /sso/ldap:
    delete:
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
      security:
      - BearerAuth: []
      summary: DeleteLDAP
      tags:
      - sso
    get:
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
      security:
      - BearerAuth: []
      summary: GetLDAP
      tags:
      - sso
    put:
      parameters:
      - description: request body
        in: body
        name: body
        schema:
          type: object
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
      security:
      - BearerAuth: []
      summary: UpsertLDAP
      tags:
      - sso
  /sso/oauth:
    get:
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
      security:
      - BearerAuth: []
      summary: ListOAuthProviders
      tags:
      - sso
  /sso/oauth/{provider}:
    delete:
      parameters:
      - description: provider
        in: path
        name: provider
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
      security:
      - BearerAuth: []
      summary: DeleteOAuthProvider
      tags:
      - sso
    put:
      parameters:
      - description: provider
        in: path
        name: provider
        required: true
        type: string
      - description: request body
        in: body
        name: body
        schema:
          type: object
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
      security:
      - BearerAuth: []
      summary: UpsertOAuthProvider
      tags:
      - sso
  /sso/oauth/{provider}/icon:
    delete:
      parameters:
      - description: provider
        in: path
        name: provider
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
      security:
      - BearerAuth: []
      summary: DeleteOAuthProviderIcon
      tags:
      - sso
    put:
      parameters:
      - description: provider
        in: path
        name: provider
        required: true
        type: string
      - description: request body
        in: body
        name: body
        schema:
          type: object
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
      security:
      - BearerAuth: []
      summary: PutOAuthProviderIcon
      tags:
      - sso
...
  /sso/saml:
    get:
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
      security:
      - BearerAuth: []
      summary: GetSAML
      tags:
      - sso
    put:
      parameters:
      - description: request body
        in: body
        name: body
        schema:
          type: object
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
      security:
      - BearerAuth: []
      summary: UpsertSAML
      tags:
      - sso
  /sso/scim:
    get:
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
      security:
      - BearerAuth: []
      summary: GetSCIM
      tags:
      - sso
    put:
      parameters:
      - description: request body
        in: body
        name: body
        schema:
          type: object
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
      security:
      - BearerAuth: []
      summary: UpsertSCIM
      tags:
      - sso
  /sso/scim/rotate-token:
    post:
      parameters:
      - description: request body
        in: body
        name: body
        schema:
          type: object
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_uigraph_app_internal_httputil.errorBody'
      security:
      - BearerAuth: []
      summary: RotateSCIMToken
      tags:
      - sso
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"